+++ src/service/implementation/ReservationServiceImpl.java
@@ -6,48 +6,48 @@ import service.ReservationService;
 import java.sql.SQLException;
 import java.sql.ResultSet;
 import java.util.ArrayList;
-import service.implementation.reservationHelper.CreateReservationHelper;
+import repository.reservationHelper.CreateReservationHelper;
 public class ReservationServiceImpl implements ReservationService {
    private ReservationRepository reservationRepository = new ReservationRepository();
     @Override
     public void addReservation(Reservation reservation) {
+        if(reservation.getBook()==null || reservation.getBorrower()==null || reservation.getDueDate()==null || reservation.getBook().getId()<0 || reservation.getBorrower().getId()<0 ){
+            System.out.println("invalid inputs!");
+            return;
+        }
         if(reservationRepository.addReservation(reservation)==true){System.out.println("Reservation added successfully!");
         } else{System.out.println("Reservation not added!");}
     }
     @Override
     public  void deleteReservation(Reservation reservation){
+        if(reservation.getId()<0){
+            System.out.println("invalid inputs!");
+            return;
+        }
         if(reservationRepository.deleteReservation(reservation)==true){System.out.println("Reservation deleted successfully!");
         }else{System.out.println("Reservation not deleted!");}
     }
     @Override
     public  List<Reservation> displayReservations(){
-        List<Reservation> reservations = new ArrayList<>();
-        try{
-            ResultSet resultSet = reservationRepository.displayReservations();
-            if(resultSet!=null){
-                while (resultSet.next()){
-                    reservations.add(CreateReservationHelper.createReservation(resultSet));
-                }
-            } else{System.out.println("No reservations found!");}
-        }catch (SQLException e){e.printStackTrace();}
+        List<Reservation> reservations = reservationRepository.displayReservations();
         return reservations;
     }
 
     @Override
     public List<Reservation> getReservationsByBorrowerCIN(String cin){
-        List<Reservation> reservations = new ArrayList<>();
-        try{
-            ResultSet resultSet = reservationRepository.getReservationsByBorrowerCIN(cin);
-            if(resultSet!=null){
-                while (resultSet.next()){
-                    reservations.add(CreateReservationHelper.createReservation(resultSet));
-                }
-            } else{System.out.println("No reservations found!");}
-        }catch (SQLException e){e.printStackTrace();}
+        if(cin==null || cin.length()==0){
+            System.out.println("invalid inputs!");
+            return null;
+        }
+        List<Reservation> reservations = reservationRepository.getReservationsByBorrowerCIN(cin);
         return reservations;
     }
     @Override
     public void changeStatusToReturned(Reservation reservation){
+         if(reservation.getId()<0){
+              System.out.println("invalid inputs!");
+              return;
+         }
        if (reservationRepository.changeStatusToReturned(reservation)==true){System.out.println("Reservation status changed successfully!");
        }else{System.out.println("Reservation status not changed!");}
      }
